chore: add UI components and fix lint issues
- Add ClaudeIcon reusable component
- Add ConfigSection component for settings pages
- Fix import ordering in task-creation hooks
- Add debug logging to task-creation service
- Remove unused cn import from api-keys settings
